// src/generated/galleryManifest.ts
export const generatedGalleryAssets: GalleryImageAsset[] = [
  {
    id: 'eb1',
    main: { webpSrc: eb1MainWebp, jpegSrc: eb1MainJpeg, width: 1600, height: 1066 },
    thumb: { webpSrc: eb1ThumbWebp, jpegSrc: eb1ThumbJpeg, width: 240, height: 160 }
  },
  {
    id: 'eb2',
    main: { webpSrc: eb2MainWebp, jpegSrc: eb2MainJpeg, width: 768, height: 1024 },
    thumb: { webpSrc: eb2ThumbWebp, jpegSrc: eb2ThumbJpeg, width: 180, height: 240 }
  },
  {
    id: 'eb3',
    main: { webpSrc: eb3MainWebp, jpegSrc: eb3MainJpeg, width: 768, height: 1024 },
    thumb: { webpSrc: eb3ThumbWebp, jpegSrc: eb3ThumbJpeg, width: 180, height: 240 }
  },
  {
    id: 'eb4',
    main: { webpSrc: eb4MainWebp, jpegSrc: eb4MainJpeg, width: 914, height: 1600 },
    thumb: { webpSrc: eb4ThumbWebp, jpegSrc: eb4ThumbJpeg, width: 137, height: 240 }
  },
  {
    id: 'eb5',
    main: { webpSrc: eb5MainWebp, jpegSrc: eb5MainJpeg, width: 1200, height: 1600 },
    thumb: { webpSrc: eb5ThumbWebp, jpegSrc: eb5ThumbJpeg, width: 180, height: 240 }
  },
  {
    id: 'eb6',
    main: { webpSrc: eb6MainWebp, jpegSrc: eb6MainJpeg, width: 1200, height: 1600 },
    thumb: { webpSrc: eb6ThumbWebp, jpegSrc: eb6ThumbJpeg, width: 180, height: 240 }
  },
  {
    id: 'eb7',
    main: { webpSrc: eb7MainWebp, jpegSrc: eb7MainJpeg, width: 1600, height: 1200 },
    thumb: { webpSrc: eb7ThumbWebp, jpegSrc: eb7ThumbJpeg, width: 240, height: 180 }
  },
  {
    id: 'eb8',
    main: { webpSrc: eb8MainWebp, jpegSrc: eb8MainJpeg, width: 1200, height: 1600 },
    thumb: { webpSrc: eb8ThumbWebp, jpegSrc: eb8ThumbJpeg, width: 180, height: 240 }
  },
  {
    id: 'eb9',
    main: { webpSrc: eb9MainWebp, jpegSrc: eb9MainJpeg, width: 900, height: 1600 },
    thumb: { webpSrc: eb9ThumbWebp, jpegSrc: eb9ThumbJpeg, width: 135, height: 240 }
  },
  {
    id: 'eb10',
    main: { webpSrc: eb10MainWebp, jpegSrc: eb10MainJpeg, width: 900, height: 1600 },
    thumb: { webpSrc: eb10ThumbWebp, jpegSrc: eb10ThumbJpeg, width: 135, height: 240 }
  },
  {
    id: 'eb11',
    main: { webpSrc: eb11MainWebp, jpegSrc: eb11MainJpeg, width: 900, height: 1600 },
    thumb: { webpSrc: eb11ThumbWebp, jpegSrc: eb11ThumbJpeg, width: 135, height: 240 }
  },
  {
    id: 'eb12',
    main: { webpSrc: eb12MainWebp, jpegSrc: eb12MainJpeg, width: 900, height: 1600 },
    thumb: { webpSrc: eb12ThumbWebp, jpegSrc: eb12ThumbJpeg, width: 135, height: 240 }
  },
  {
    id: 'eb13',
    main: { webpSrc: eb13MainWebp, jpegSrc: eb13MainJpeg, width: 900, height: 1600 },
    thumb: { webpSrc: eb13ThumbWebp, jpegSrc: eb13ThumbJpeg, width: 135, height: 240 }
  },
  {
    id: 'eb14',
    main: { webpSrc: eb14MainWebp, jpegSrc: eb14MainJpeg, width: 900, height: 1600 },
    thumb: { webpSrc: eb14ThumbWebp, jpegSrc: eb14ThumbJpeg, width: 135, height: 240 }
  },
  {
    id: 'eb15',
    main: { webpSrc: eb15MainWebp, jpegSrc: eb15MainJpeg, width: 1600, height: 1066 },
    thumb: { webpSrc: eb15ThumbWebp, jpegSrc: eb15ThumbJpeg, width: 240, height: 160 }
  },
  {
    id: 'eb16',
    main: { webpSrc: eb16MainWebp, jpegSrc: eb16MainJpeg, width: 1200, height: 1600 },
    thumb: { webpSrc: eb16ThumbWebp, jpegSrc: eb16ThumbJpeg, width: 180, height: 240 }
  },
  {
    id: 'eb17',
    main: { webpSrc: eb17MainWebp, jpegSrc: eb17MainJpeg, width: 537, height: 1600 },
    thumb: { webpSrc: eb17ThumbWebp, jpegSrc: eb17ThumbJpeg, width: 80, height: 240 }
  },
  {
    id: 'eb18',
    main: { webpSrc: eb18MainWebp, jpegSrc: eb18MainJpeg, width: 1600, height: 1200 },
    thumb: { webpSrc: eb18ThumbWebp, jpegSrc: eb18ThumbJpeg, width: 240, height: 180 }
  },
  {
    id: 'eb19',
    main: { webpSrc: eb19MainWebp, jpegSrc: eb19MainJpeg, width: 1179, height: 1594 },
    thumb: { webpSrc: eb19ThumbWebp, jpegSrc: eb19ThumbJpeg, width: 177, height: 240 }
  },
  {
    id: 'eb20',
    main: { webpSrc: eb20MainWebp, jpegSrc: eb20MainJpeg, width: 1572, height: 1600 },
    thumb: { webpSrc: eb20ThumbWebp, jpegSrc: eb20ThumbJpeg, width: 236, height: 240 }
  },
  {
    id: 'eb21',
    main: { webpSrc: eb21MainWebp, jpegSrc: eb21MainJpeg, width: 1600, height: 1298 },
    thumb: { webpSrc: eb21ThumbWebp, jpegSrc: eb21ThumbJpeg, width: 240, height: 194 }
  }
];